refactor(cli): lazy-load command handlers via dynamic import
- Add lazyHandler utility matching new-ui pattern
- Split each command into command.ts (descriptor) + run.ts (handler)
- Restructure flat files into directory-per-command convention
- Defer heavy deps (viem, graphql, chalk, sablier) until execution

Closes #290
