+++ cli/commands/check-vendors/command.ts
@@ -1,0 +1,17 @@
+import { Command, Options } from "@effect/cli";
+import { lazyHandler } from "../../utils/lazy-command.js";
+
+const chainIdOption = Options.integer("chain-id").pipe(
+  Options.withDescription("Chain ID to check"),
+  Options.withAlias("c")
+);
+
+export const checkVendorsCommand = Command.make(
+  "check-vendors",
+  { chainId: chainIdOption },
+  ({ chainId }) =>
+    lazyHandler(
+      () => import("./run.js"),
+      (m) => m.handler({ chainId })
+    )
+);
