+++ cli/commands/prices/check/command.ts
@@ -1,0 +1,9 @@
+import { Command } from "@effect/cli";
+import { lazyHandler } from "../../../utils/lazy-command.js";
+
+export const pricesCheckCommand = Command.make("prices-check", {}, () =>
+  lazyHandler(
+    () => import("./run.js"),
+    (m) => m.handler()
+  )
+);
