+++ cli/commands/prices/sync/run.ts
@@ -1,13 +1,12 @@
 import * as path from "node:path";
-import { Command as CliCommand } from "@effect/cli";
 import { FileSystem } from "@effect/platform";
 import chalk from "chalk";
 import { Console, Effect } from "effect";
-import { colors, createTable, displayHeader } from "../../display.js";
-import { ProcessError } from "../../errors.js";
-import type { PriceDataFile } from "../../price-data.js";
-import { getRequiredPriceDataFiles } from "../../price-data.js";
-import { CliEnv } from "../../services/env.js";
+import { colors, createTable, displayHeader } from "../../../display.js";
+import { ProcessError } from "../../../errors.js";
+import type { PriceDataFile } from "../../../price-data.js";
+import { getRequiredPriceDataFiles } from "../../../price-data.js";
+import { CliEnv } from "../../../services/env.js";
 
 /**
  * Get list of required TSV files based on Sablier chain data and forex rates.
@@ -30,7 +29,7 @@ type SyncResult = {
   status: "copied" | "error" | "missing";
 };
 
-const priceDataSyncLogic = () =>
+export const handler = () =>
   Effect.gen(function* () {
     const env = yield* CliEnv;
     const fs = yield* FileSystem.FileSystem;
@@ -145,5 +144,3 @@ const priceDataSyncLogic = () =>
       );
     }
   });
-
-export const pricesSyncCommand = CliCommand.make("prices-sync", {}, priceDataSyncLogic);
