+++ cli/index.ts
@@ -10,24 +10,24 @@
 import { Command } from "@effect/cli";
 import { NodeRuntime } from "@effect/platform-node";
 import { Effect } from "effect";
-import { checkVendorsCommand } from "./commands/check-vendors.js";
-import { envioConfigCommand } from "./commands/codegen/envio-config.js";
-import { graphManifestCommand } from "./commands/codegen/graph-manifest.js";
-import { schemaCommand } from "./commands/codegen/schema.js";
-import { exportSchemaCommand } from "./commands/export-schema.js";
-import { graphDeployAllCommand } from "./commands/graph/deploy/all.js";
-import { graphDeploySingleCommand } from "./commands/graph/deploy/single.js";
+import { checkVendorsCommand } from "./commands/check-vendors/command.js";
+import { envioConfigCommand } from "./commands/codegen/envio-config/command.js";
+import { graphManifestCommand } from "./commands/codegen/graph-manifest/command.js";
+import { schemaCommand } from "./commands/codegen/schema/command.js";
+import { exportSchemaCommand } from "./commands/export-schema/command.js";
+import { graphDeployAllCommand } from "./commands/graph/deploy/all/command.js";
+import { graphDeploySingleCommand } from "./commands/graph/deploy/single/command.js";
 import { graphOpenStudioCommand } from "./commands/graph/open-studio.js";
-import { graphReviveCommand } from "./commands/graph/revive.js";
-import { pricesCheckCommand } from "./commands/prices/check.js";
-import { pricesSyncCommand } from "./commands/prices/sync.js";
-import { printChainsCommand } from "./commands/print-chains.js";
-import { queryActionsCommand } from "./commands/query/actions.js";
-import { queryAssetsCommand } from "./commands/query/assets.js";
-import { queryAverageMauCommand } from "./commands/query/average-mau.js";
-import { queryTotalUsdFeesCommand } from "./commands/query/total-usd-fees.js";
-import { queryUniqueTxsCommand } from "./commands/query/unique-txs.js";
-import { recoverTokensCommand } from "./commands/recover-tokens.js";
+import { graphReviveCommand } from "./commands/graph/revive/command.js";
+import { pricesCheckCommand } from "./commands/prices/check/command.js";
+import { pricesSyncCommand } from "./commands/prices/sync/command.js";
+import { printChainsCommand } from "./commands/print-chains/command.js";
+import { queryActionsCommand } from "./commands/query/actions/command.js";
+import { queryAssetsCommand } from "./commands/query/assets/command.js";
+import { queryAverageMauCommand } from "./commands/query/average-mau/command.js";
+import { queryTotalUsdFeesCommand } from "./commands/query/total-usd-fees/command.js";
+import { queryUniqueTxsCommand } from "./commands/query/unique-txs/command.js";
+import { recoverTokensCommand } from "./commands/recover-tokens/command.js";
 import { CliLive } from "./context.js";
 import { CliEnv } from "./services/env.js";
